Backend/requirements.txt
django
djangorestframework
markdown
djangorestframework-simplejwt